Get rid of `quickstart` term in the `web/server` module.
