+++ web/server/src/main/java/io/spine/tasks/web/CorsFilter.java
@@ -18,7 +18,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.quickstart.tasks.web;
+package io.spine.tasks.web;
 
 import com.google.common.net.HttpHeaders;
